improved file structure for easier traversal
