+++ src/Pages/TeamToTeam/NHL/components/TeamStats.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { singleNhlTeamRetrieval } from '../../../api/nhlApi';
+import { singleNhlTeamRetrieval } from '../../../../api/nhlApi';
 import SeasonFilter from './SeasonFilter';
 
-export default function TeamStats({setCurrentTeam, id }) {
+export default function TeamStats({ setCurrentTeam, id }) {
   const [stats, setStats] = useState({});
   const [seasonStats, setSeasonStats] = useState({});
   const [seasonRank, setSeasonRank] = useState({});
@@ -33,10 +33,18 @@ export default function TeamStats({setCurrentTeam, id }) {
     'Overtime (losses)': `${seasonStats.ot} (${seasonRank.ot})`,
     'Total Points': `${seasonStats.pts} (${seasonRank.pts})`,
     'Faceoff Win %': `${seasonStats.faceOffWinPercentage} (${seasonRank.faceOffWinPercentage})`,
-    'Goals Per Game': `${seasonStats.goalsPerGame.toFixed(1)} (${seasonRank.goalsPerGame})`,
-    'Goals Against Per Game': `${seasonStats.goalsAgainstPerGame.toFixed(1)} (${seasonRank.goalsAgainstPerGame})`,
-    'Shots Per Game': `${seasonStats.shotsPerGame.toFixed(1)} (${seasonRank.shotsPerGame})`,
-    'Shots Allowed Per Game': `${seasonStats.shotsAllowed.toFixed(1)} (${seasonRank.shotsAllowed})`,
+    'Goals Per Game': `${seasonStats.goalsPerGame.toFixed(1)} (${
+      seasonRank.goalsPerGame
+    })`,
+    'Goals Against Per Game': `${seasonStats.goalsAgainstPerGame.toFixed(1)} (${
+      seasonRank.goalsAgainstPerGame
+    })`,
+    'Shots Per Game': `${seasonStats.shotsPerGame.toFixed(1)} (${
+      seasonRank.shotsPerGame
+    })`,
+    'Shots Allowed Per Game': `${seasonStats.shotsAllowed.toFixed(1)} (${
+      seasonRank.shotsAllowed
+    })`,
     'Power Plays Per Game': `${seasonStats.powerPlayOpportunities} (${seasonRank.powerPlayOpportunities})`,
     'Power Play %': `${seasonStats.powerPlayPercentage} (${seasonRank.powerPlayPercentage})`,
     'Penalty Kill %': `${seasonStats.penaltyKillPercentage} (${seasonRank.penaltyKillPercentage})`,
@@ -44,7 +52,11 @@ export default function TeamStats({setCurrentTeam, id }) {
 
   return (
     <div className='each-team-stats'>
-      <SeasonFilter setSeason={setSeason} season={season} startDate={stats.firstYearOfPlay} />
+      <SeasonFilter
+        setSeason={setSeason}
+        season={season}
+        startDate={stats.firstYearOfPlay}
+      />
       <img
         src={`https://www-league.nhlstatic.com/images/logos/teams-current-primary-light/${id}.svg`}
         alt={`Hello`}
@@ -76,13 +88,13 @@ export default function TeamStats({setCurrentTeam, id }) {
         </tbody>
       </table>
       <button
-            className='player-text'
-            onClick={() => {
-              setCurrentTeam('');
-            }}
-          >
-            Search
-          </button>
+        className='player-text'
+        onClick={() => {
+          setCurrentTeam('');
+        }}
+      >
+        Search
+      </button>
     </div>
   );
 }
